sdks/blob-az/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blob-az"
version = "0.2.0"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Async Python SDK for Azure Blob Storage (both block and append blobs)"
dependencies = [
    "azure-storage-blob", "pydantic", "aiohttp", "haskellian"    
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.urls]
repo = "https://github.com/moveread/azure-sdks"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]